+++ utils/fight-systems-check.js
@@ -29,6 +29,7 @@ const FILES = [
   "src/core/fishing/hold_opposition_resolver.js",
   "src/core/fishing/fish_direction_intent_sampler.js",
   "src/core/fishing/fish_fight_direction_resolver.js",
+  "src/core/fishing/endurance/endurance_movement_debuff_calculator.js",
   "src/core/fishing/drag_force_calculator.js",
   "src/core/fishing/line_radial_movement_splitter.js",
   "src/core/fishing/landing_lift_tension_calculator.js",